vault33-gatekeeper/requirements.txt
discord.py==2.4.0
python-telegram-bot==21.6
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.0
pydantic==2.9.2
python-dotenv==1.0.1
xrpl-py==3.0.0
cryptography>=42.0.8
httpx==0.27.2
loguru==0.7.2
python-multipart>=0.0.9